bug(feedback): implement feedback 2 [finshes #165414209]
