+++ .gitignore
@@ -1,3 +1,3 @@
+.vscode/
 node_modules/
 .nyc_output/
-.vscode/
