+++ Backend/routes/account.route.js
@@ -1,7 +1,8 @@
 import express from 'express';
-import Account from '../services/account.service';
+import Account from '../controllers/account.controller';
+import AccountMiddleware from '../middleware/account.middleware';
 
 const router = express.Router();
 
-router.patch('/:accountNumber', Account.checkAuthorization, Account.checkStaffAccess, Account.changeAccountStatus);
+router.patch('/:accountNumber', AccountMiddleware.checkAuthorization, AccountMiddleware.checkStaffAccess, Account.changeAccountStatus);
 export default router;
