+++ Backend/routes/auth.route.js
@@ -1,11 +1,11 @@
 import express from 'express';
-import Validate from '../services/signup.service';
-import User from '../controllers/signup.controller';
-import Signin from '../services/signin.service';
-import Util from '../services/util.service';
+import SigninMiddleware from '../middleware/signin.middleware';
+import signupController from '../controllers/signup.controller';
+import SignupMiddleware from '../middleware/signup.middleware';
+import Util from '../helper/util.helper';
 
 const router = express.Router();
 
-router.post('/signup', Validate.checkEmptyFields, Validate.checkValidInput, User.create);
-router.post('/signin', Signin.checkEmptyFields, Signin.checkExistence, Signin.getId, Util.sendToken);
+router.post('/signup', SignupMiddleware.checkEmptyFields, SignupMiddleware.checkValidInput, signupController.createUser);
+router.post('/signin', SigninMiddleware.checkEmptyFields, SigninMiddleware.checkExistence, SigninMiddleware.getId, Util.sendToken);
 export default router;
