+++ UI/admin-createUser.html
@@ -48,7 +48,7 @@
           <input type="text" name="firstname" class="top-input" placeholder="First name" required autofocus><br>
           <input type="text" name="lastName" placeholder="Last name" required><br>
           <input type="email" name="email" placeholder="Email" required><br>
-          <input type="password" name="password" placeholder="Password" required><br>
+          <input type="password" name="password" placeholder="Password" required><br><br>
           <label for="userType">User type:</label>
           <select class="user-type" id="userType" name="userType">
             <option value="staff">Staff</option>
