+++ UI/admin-dashboard.html
@@ -68,7 +68,7 @@
           <tr class="trans-history">
             <td>3456355123</td>
             <td>Martha Stewert</td>
-            <td>savings</td>
+            <td>Savings</td>
             <td>07/12/2005</td>
             <td><a href="javascript:void(0);" onclick="openModal()"><i class="fas fa-external-link-alt"></i></a></td>
           </tr>
@@ -82,7 +82,7 @@
           <tr class="trans-history">
             <td>3056749900</td>
             <td>Jaime Fraser</td>
-            <td>savings</td>
+            <td>Savings</td>
             <td>24/08/2012</td>
             <td><a href="javascript:void(0);" onclick="openModal()"><i class="fas fa-external-link-alt"></i></a></td>
           </tr>
@@ -115,7 +115,7 @@
           <tr class="trans-history">
             <td>3456355123</td>
             <td>Martha Stewert</td>
-            <td>savings</td>
+            <td>Savings</td>
             <td><a href="javascript:void(0);" onclick="openModal()"><i class="fas fa-external-link-alt"></i></a></td>
           </tr>
           <tr class="trans-history">
@@ -127,7 +127,7 @@
           <tr class="trans-history">
             <td>3056749900</td>
             <td>Jaime Fraser</td>
-            <td>savings</td>
+            <td>Savings</td>
             <td><a href="javascript:void(0);" onclick="openModal()"><i class="fas fa-external-link-alt"></i></a></td>
           </tr>
           <tr class="trans-history">
