+++ UI/admin-editDetails.html
@@ -53,7 +53,7 @@
             <option value="No">No</option>
             <option value="Yes">Yes</option>
           </select><br><br>
-          <input type="password" name="Oldpassword" id="old-pswd" class="pass" placeholder="Old Password" required autofocus><br>
+          <input type="password" name="Oldpassword" id="old-pswd" class="pass top-input" placeholder="Old Password" required autofocus>
           <input type="password" name="Newpassword" id="new-pswd" class='bottom-input pass' placeholder="New password" required><br>
           <input type="submit" value="Submit">
         </form>
