+++ UI/create-account.html
@@ -18,10 +18,10 @@
     </header>
     <form class="signupForm small-form" action="./account-dashboard.html" method="post">
       <h4>Create bank account</h4>
-      <input type="text" name="accountName" class="top-input" placeholder="Account name" required autofocus>
-      <input type="email" name="email" placeholder="Email" required><br>
+      <input type="text" name="accountName" class="top-input" value="Williams Smith"  disabled>
+      <input type="email" name="email" placeholder="Email" value="willi@yahoo.com" required><br>
       <input type="tel" name="mobileNumber" placeholder="Phone number" required><br>
-      <input type="text" name="address" class="bottom-input" placeholder="Address" required><br>
+      <input type="text" name="address" class="bottom-input" placeholder="Address" required><br><br>
       <label for="userType">Account Type:</label>
       <select class="account-type" id="type" name="accountType">
         <option value="current">Current</option>
