+++ UI/staff-accountDashboard.html
@@ -65,7 +65,7 @@
           <tr class="trans-history">
             <td>3456355123</td>
             <td>Martha Stewert</td>
-            <td>savings</td>
+            <td>Savings</td>
             <td>07/12/2005</td>
             <td><a href="javascript:void(0);" onclick="openModal()"><i class="fas fa-external-link-alt"></i></a></td>
           </tr>
@@ -79,7 +79,7 @@
           <tr class="trans-history">
             <td>3056749900</td>
             <td>Jaime Fraser</td>
-            <td>savings</td>
+            <td>Savings</td>
             <td>24/08/2012</td>
             <td><a href="javascript:void(0);" onclick="openModal()"><i class="fas fa-external-link-alt"></i></a></td>
           </tr>
@@ -112,7 +112,7 @@
           <tr class="trans-history">
             <td>3456355123</td>
             <td>Martha Stewert</td>
-            <td>savings</td>
+            <td>Savings</td>
             <td><a href="javascript:void(0);" onclick="openModal()"><i class="fas fa-external-link-alt"></i></a></td>
           </tr>
           <tr class="trans-history">
@@ -124,7 +124,7 @@
           <tr class="trans-history">
             <td>3056749900</td>
             <td>Jaime Fraser</td>
-            <td>savings</td>
+            <td>Savings</td>
             <td><a href="javascript:void(0);" onclick="openModal()"><i class="fas fa-external-link-alt"></i></a></td>
           </tr>
           <tr class="trans-history">
@@ -147,7 +147,7 @@
           <p class='title'>Account Type:</p>
           <p>Current</p>
           <p class='title'>Account Status:</p>
-          <p>Active/p>
+          <p>Active</p>
           <p class='title'>Account Balance:</p>
           <p>#30000</p>
           <p class='title'>Date Created:</p>
