+++ UI/transaction-history.html
@@ -102,19 +102,13 @@
               <tr class="trans-history">
                 <td>30/05/2017</td>
                 <td>#3000</td>
-                <td>dbt</td>
+                <td>Dbt</td>
                 <td><a href="javascript:void(0);" onclick="openModal()"><i class="fas fa-external-link-alt"></i></a></td>
               </tr>
               <tr class="trans-history">
                 <td>30/05/2017</td>
                 <td>#3000</td>
-                <td>dbt</td>
-                <td><a href="javascript:void(0);" onclick="openModal()"><i class="fas fa-external-link-alt"></i></a></td>
-              </tr>
-              <tr class="trans-history">
-                <td>30/05/2017</td>
-                <td>#3000</td>
-                <td>Cr</td>
+                <td>Dbt</td>
                 <td><a href="javascript:void(0);" onclick="openModal()"><i class="fas fa-external-link-alt"></i></a></td>
               </tr>
               <tr class="trans-history">
@@ -126,13 +120,19 @@
               <tr class="trans-history">
                 <td>30/05/2017</td>
                 <td>#3000</td>
-                <td>dbt</td>
+                <td>Cr</td>
                 <td><a href="javascript:void(0);" onclick="openModal()"><i class="fas fa-external-link-alt"></i></a></td>
               </tr>
               <tr class="trans-history">
                 <td>30/05/2017</td>
                 <td>#3000</td>
-                <td>dbt</td>
+                <td>Dbt</td>
+                <td><a href="javascript:void(0);" onclick="openModal()"><i class="fas fa-external-link-alt"></i></a></td>
+              </tr>
+              <tr class="trans-history">
+                <td>30/05/2017</td>
+                <td>#3000</td>
+                <td>Dbt</td>
                 <td><a href="javascript:void(0);" onclick="openModal()"><i class="fas fa-external-link-alt"></i></a></td>
               </tr>
             </table>
@@ -143,16 +143,20 @@
         <a href="javascript:void(0);" class="close-icon" onclick="closeModal()"><i class="far fa-minus-square"></i></a>
         <div class="modal-head"></div>
         <div class="modal-body">
-            <p class="title">Amount:</p>
-            <p>#3000</p>
             <p class="title">Transaction Date:</p>
             <p>30/05/2017</p>
             <p class="title">Account Number:</p>
             <p>3077145675</p>
-            <p class="title">Narration:</p>
-            <p>Atm</p>
             <p class="title">Transaction Type:</p>
             <p>Credit</p>
+            <p class="title">Amount:</p>
+            <p>#3000</p>
+            <p class="title">Narration:</p>
+            <p>Atm</p>
+            <p class="title">Old Balance:</p>
+            <p>#3000</p>
+            <p class="title">New Balance:</p>
+            <p>#6000</p>
         </div>
       </div>
     </div>
